rename local dev format
Renaming the local development format allows a user to install both the dev version of the format and a remote stable version. This makes it possible to toggle back and forth between versions and see if you've broken osmething.

IIUC this should not affect the build process as webpack reads `template/format.js` instead of `src/format-dev.js`.